+++ code/modules/economy/Economy_Events.dm
@@ -94,12 +94,10 @@
 			if(FESTIVAL)
 				newMsg.body = "A [pick("festival","week long celebration","day of revelry","planet-wide holiday")] has been declared on [affected_dest.name] by [pick("Governor","Commissioner","General","Commandant","Administrator")] [random_name(pick(MALE,FEMALE))] to celebrate [pick("the birth of their [pick("son","daughter")]","coming of age of their [pick("son","daughter")]","the pacification of rogue military cell","the apprehension of a violent criminal who had been terrorising the planet")]. Massive stocks of food and meat have been bought driving up prices across the planet."
 
-	for(var/datum/feed_channel/FC in GLOB.news_network.channels)
-		if(FC.channel_name == "Nyx Daily")
-			FC.add_message(newMsg)
-			break
-	for(var/obj/machinery/newscaster/NEWSCASTER in GLOB.allNewscasters)
-		NEWSCASTER.alert_news("Nyx Daily")
+	GLOB.news_network.get_channel_by_name("Nyx Daily")?.add_message(newMsg)
+	for(var/nc in GLOB.allNewscasters)
+		var/obj/machinery/newscaster/NC = nc
+		NC.alert_news("Nyx Daily")
 
 /datum/event/economic_event/end()
 	for(var/good_type in dearer_goods)
